+++ SourceCode/Services/Extensions/LayoutParticipantExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace ModulesRegistry.Services.Extensions;
+
+public static class LayoutParticipantExtensions
+{
+    public static string NameWithCityAndCountry(this LayoutParticipant? layoutParticipant) =>
+        layoutParticipant is null || layoutParticipant.Person is null ? string.Empty :
+        layoutParticipant.Person.NameWithCityAndCountry();
+
+}